% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aligning_FT_QTOF_SQL.R
\name{Aligning_FT_QTOF_SQL}
\alias{Aligning_FT_QTOF_SQL}
\title{Aligning FTMS and QTOF experiments with the same/different polarities}
\usage{
Aligning_FT_QTOF_SQL(
  FT_path,
  QTOF_path,
  FT_expnr,
  QTOF_expnr,
  Assoc = NULL,
  err = 0.02,
  t.ini = 5,
  lc.err = 1,
  rng = 2,
  minIon = 0.6,
  startpoint = 1,
  save_assoc = FALSE,
  aggregated_Ft = FALSE,
  aggregated_QTOF = FALSE
)
}
\arguments{
\item{FT_path}{\code{character(1)} the path to the FTMS SQL database.}

\item{QTOF_path}{\code{character(1)} the path to the QTOF SQL database.}

\item{FT_expnr}{\code{numeric(1)} the FTMS reference experiment to align.}

\item{QTOF_expnr}{\code{numeric(1)} the QTOF experiment to align with the previous
FTMS experiment.}

\item{Assoc}{\code{character(1)} the file that contains some previous alignment
results done by the user. If \code{NULL} we create it in the function.}

\item{err}{\code{numeric(1)} is the m/z tolerance window used to decide which QTOF
peaks are potential isomeric matches for a given FT peak.}

\item{t.ini}{\code{numeric(1)} used in the \code{Aligning_General_SQL()} to define the
number of neighboring peaks.}

\item{lc.err}{\code{numeric(1)} the retention-time alignment tolerance.}

\item{rng}{\code{numeric(1)} used in \code{RemoveOutliers()} refers to the number of
standard  deviations used to define the outlier cutoff.}

\item{minIon}{\code{numeric(1)} Minimum number of matched peaks required to
consider a valid match. Matches with fewer than \code{minIon} shared
ions are discarded.}

\item{startpoint}{\code{numeric(1)} used in \code{RegressionPie_LCalign()} refers to
the minimum retention time for the first knot in piecewise regression.}

\item{save_assoc}{\code{logical(1)} If true we add the alignment result to the
Assoc text file.}

\item{aggregated_Ft}{\code{logical(1)} If true the use of the aggregated ftms
spectra of all levels is involved.}

\item{aggregated_QTOF}{\code{logical(1)} If true the use of the aggregated QTOF
spectra of all collision energies is involved.}
}
\value{
A \code{data.frame} containing the matched FT–QTOF compounds with columns:
\itemize{
\item \code{ref_compid}: FTMS compound ID
\item \code{target_compid}: QTOF compound ID
\item \code{ref_database}: filename of the FTMS database
\item \code{target_database}: filename of the QTOF database.
This table is updated with new associations found by the alignment procedure.
}
}
\description{
\code{Aligning_FT_QTOF_SQL()} aligns FTMS and QTOF experiments with the same
or different polarities, by calling a set of defined functions, and adds the
aligned compounds to a given text file.
}
\details{
the function reads the both sql databases FTMS and QTOF, and extracts
their mode, create an output file to store the alignment results in a .txt
file if it's already given by the user otherwise we create a new one.
Then calls a set of defined functions :
\itemize{
\item \code{Aligning_General_SQL()}: performs the local alignment, and adjust the mz
values if the alignment is between two different polarities.
\item \verb{matchFTSyn_SQL(): removes the matches found by the previous function if they contain fewer than }minIon` matched peaks.
\item \code{RemoveOutliers()}: calculate the standard deviation in the rt difference
found between each matched feature and remove all rows outside a range.
\item \code{RegressionPie_LCalign()}: apply a regression model  to reduce the
influence of false matches and outliers.
\item \code{FillAssocFTnQTOFn_SQL()}: store the final aligned compounds to a .txt
file.
}
}
\author{
Ahlam Mentag
}
