% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TablewiseIdent.R
\name{TablewiseIdent}
\alias{TablewiseIdent}
\title{0' because the unknown got the name from a CSPP present}
\usage{
TablewiseIdent(base.dir, finlist, SubDB, nOc = NULL)
}
\description{
0' because the unknown got the name from a CSPP present
}
